Fix agenda due-date issues (#41)
* org-agenda: actually show today’s entries

In the docstring, `dashboard-date-due-p` claims that it will check whether
`timestamp` is today or in the past. However, the function will actually only
return `t` if timestamp *precedes* today. Since this predicate is used to filter
agenda entries, entries which are due today are not actually shown.

We now also show items scheduled for today by checking whether the deadline or
scheduled date precedes *tomorrow*.

* org-agenda: fix logic error when filtering due entries

We need to properly guard the invocation of `dashboard-date-due-p` against nil,
since `decode-time` will otherwise (unwantedly) fill in the current time as a
due date.