Fix dashboard-setup-startup-hook
Hooks from `emacs-startup-hook` are called too early so dashboard
incorrectly drawn text version of logo on some occasions. Steps to
reproduce:

* Start `emacs --daemon` from terminal
* Open `emacsclient -t`
* Open `emacsclient -c` -- text version of logo is displayed

This behavior is fixed by utilizing `window-setup-hook`. Emacs'
documentation of `window-setup-hook`:

This normal hook is very similar to emacs-startup-hook. The only
difference is that it runs slightly later, after setting of the frame
parameters.