Fix error when calling dashboard-previous-section before any section
Place cursor on "Welcome to Emacs!" or just anywhere before the first
section, then run `dashboard-previous-section` --- it will error out
(at goto-char) complaining that nil is not a number or marker. This is
caused by the current version not checking if local variables
previous-section-start or current-section-start are non-nil (unlike
dashboard-next-section).