Bugfix dashboard-filter-agenda-by-time.
When schedule-time or deadline-time is nil but not both,
org-time-less-p treat them as current-date. Filter will skip the entry
unless it's not done and has schedule time less than due date or has a
deadline time less than due date.

There are other ways to express this condition
```ruby
(when (or (org-entry-is-done-p)
          (and (or (null schedule-time)
                   (org-time-less-p due-date schedule-time))
               (or (null deadline-time)
                   (org-time-less-p due-date deadline-time))))
  (point))
```
Or
```ruby
(cond
  ((org-entry-is-done-p) (point))
  ((and schedule-time
        (org-time-less-p schedule-time due-date)) nil)
  ((and deadline-time
        (org-time-less-p deadline-time due-date)) nil)
  (t (point)))
```