Agenda highlight headlines
Introduces `dashboard-agenda-highlight-headlines` as a Boolean to
highlight agenda headlines using org default faces.

Headlines are highlighted by setting the corresponding
`face` text-property to each part of the headline. Issue #264 shows
that headlines are highlighted only after visit the org buffer and
when org buffer is killed headlines looses their faces.

I couldn't found the function that sets the faces, so now
faces are assigned to the TODO and HEADING parts only if they don't
already had a face.